# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Merlin Route Finder API
  description: >
    Contrato da API REST do Merlin Route Finder (FIAP Challenge 2026 / Leroy Merlin).
    Escrito à mão como fonte de verdade API-First (ver docs/planejamento-tecnico.md),
    antes da implementação dos controllers reais. Cobre os recursos das Fases 0 a 3
    do backlog revisado (docs/backlog-trello-revisado.md).
  version: "0.1.0"
servers:
  - url: /api/v1
...
paths:
...
  /roteiro/itens/{itemId}/coletar:
    patch:
      tags: [Roteiro]
      summary: Marcar item como coletado durante a caminhada (UC-014)
      description: >
        Acionado pelo Mobile quando o cliente confirma ter pego um produto. É o mecanismo
        que faz o "ponto" avançar visualmente na simulação de caminhada do Mobile PWA
        (estratégia de demo da banca final).
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: Item marcado como coletado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemRoteiroDetalhado"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /roteiro/itens/{itemId}/ruptura:
    post:
      tags: [RupturaEstoque]
      summary: Tratar ruptura de estoque e sugerir substituto (UC-013)
      description: >
        Acionado quando o cliente sinaliza "Prateleira Vazia" no Mobile. Registra a
        ruptura, faz uma pré-filtragem espacial por corredores adjacentes ao PontoMapa
        do produto em falta, e aciona o Gemini via function calling para eleger o
        substituto semântico mais adequado dentre os candidatos já filtrados
        espacialmente (nunca por texto livre do modelo).
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: Substituto sugerido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RupturaEstoqueResponse"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/SessaoInativa"
        "422":
          description: >
            Ruptura registrada, mas nenhum substituto plausível encontrado nas
            proximidades - seja porque nenhum produto disponível está dentro do raio,
            seja porque o assistente avaliou os candidatos e concluiu que nenhum cumpre
            a mesma função.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
...
components:
  parameters:
...
    ItemId:
      name: itemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    NaoEncontrado:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StandardError"
    SessaoInativa:
      description: >
        A sessão existe mas não está mais ativa (encerrada, expirada ou abandonada), então a
        operação não é permitida. Diferente do 404: o recurso existe, o estado é que impede.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StandardError"
...
  schemas:
...
    TipoPonto:
      type: string
      enum: [PRATELEIRA, BANHEIRO, CAIXA, TOTEM]
...
    PontoMapa:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tipo:
          $ref: "#/components/schemas/TipoPonto"
        corredor:
          type: string
        coordenadaX:
          type: integer
        coordenadaY:
          type: integer
    Produto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        nome:
          type: string
        preco:
          type: number
          format: double
        saldoEstoque:
          type: integer
        pontoMapaId:
          type: string
          format: uuid
    ProdutoDetalhado:
      allOf:
        - $ref: "#/components/schemas/Produto"
        - type: object
          properties:
            pontoMapa:
              $ref: "#/components/schemas/PontoMapa"
...
    ItemRoteiro:
      type: object
      properties:
        id:
          type: string
          format: uuid
        produtoId:
          type: string
          format: uuid
        ordemCaminho:
          type: integer
        coletado:
          type: boolean
    ItemRoteiroDetalhado:
      allOf:
        - $ref: "#/components/schemas/ItemRoteiro"
        - type: object
          properties:
            produto:
              $ref: "#/components/schemas/Produto"
...
    RupturaEstoqueResponse:
      type: object
      properties:
        produtoOriginalId:
          type: string
          format: uuid
        produtoSugeridoId:
          type: string
          format: uuid
        produtoSugerido:
          $ref: "#/components/schemas/ProdutoDetalhado"
        justificativa:
          type: string
          description: Explicação em linguagem natural para a sugestão
        origemSugestao:
          type: string
          description: >
            Quem elegeu o substituto. ASSISTENTE_IA quando o modelo escolheu entre os
            candidatos pré-filtrados; PROXIMIDADE quando o assistente estava indisponível
            e o sistema caiu para o produto disponível mais próximo, sem análise semântica.
            O frontend deve usar este campo para não rotular como recomendação inteligente
            o que foi apenas o item mais perto.
          enum: [ASSISTENTE_IA, PROXIMIDADE]
...
    StandardError:
      type: object
      description: Espelha br.com.jence.backend.presentation.response.StandardError
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        validationErrors:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string